+++ CommonInterfaces/Interfaces/IConnectionService.cs
@@ -5,7 +5,5 @@ namespace CommonInterfaces
     public interface IConnectionService
     {
         public Task<IUser?> ReceieveMessage(int id);
-        public void SendToMiners(string data);
-        
     }
 }
