+++ CommonInterfaces/Interfaces/IReceiver.cs
@@ -1,0 +1,7 @@
+namespace CommonInterfaces 
+{
+    public interface IListReceiver 
+    {
+        Task<List<Miner>> Receive();
+    }
+}
