+++ Miner/Program.cs
@@ -7,11 +7,12 @@ internal class Program
     private async static Task Main(string[] args)
     {
         var miner = new Miner(); 
+        var receiver = new MinerReceivingService();
         await miner.Register();
         Console.WriteLine("Miner registered successfuly!");
         Console.WriteLine(miner);
 
-        var uiHandler = new MinerUiHandler(miner);
+        var uiHandler = new MinerUiHandler(receiver);
         await uiHandler.HandleUI();
     }
 }
